mobile sites: Do not ever expire peer addresses
For mobile sites, peer addresses come from our config or DNS name
lookup.  Ones that are not working now may work later in a different
network environment.  The mobile end is in charge of public path
selection so it needs to retain the information to try these
currently-non-working addresses.

In practice, this change allows me to switch backwards and forwards
between the FOSDEM v6-only wifi, and my v4-only USB stick.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>