make-secnet-sites: Tolerate unused group in additions to location
When processing a sites file other than via userv, the group (if
specified) is not of any interest.

But since tainting, we need to bless it for re-output.  (This is
necessary even in modes where we don't actually write anything out,
like .conf generation.)

The correct place to do this is in OpConf's base class.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>