make-secnet-sites: Tolerate missing group in userv sites file
When processing a sites file fragment via userv, the group in the
provided file is supposed to be optional.  This was accidentally
regressed in 19482a2958fa
   make-secnet-sites: Do not write out unchecked output in sites

The additional call to w[2].groupname ought to have been conditional.
Now we have precisely the right code in OpBase, so simply call it.

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>