autoconf: Check for adns_if_monotonic and define USE_MONOTONIC
Nothing uses this yet.  Split out into its own commit for clarity.

This flag is not yet in any released version of adns but as adns
upstream I promise this is how it will be.  for my convenience I am
making secnet check for it ahead of its existence in public adns....

Signed-off-by: Ian Jackson <ijackson@chiark.greenend.org.uk>